* '1.0.0' of github.com:threefoldfoundation/app_backend:
  Fix check if referrer was already set or not
  Decline users from connecting to the server from other apps then threefold
  Round investment agreement amount and store it as float

+++ plugins/tff_backend/models/investor.py
@@ -14,7 +14,7 @@ class InvestmentAgreement(NdbModel):
     STATUS_PAID = 2
 
     app_user = ndb.UserProperty()
-    amount = ndb.IntegerProperty(indexed=False)
+    amount = ndb.FloatProperty(indexed=False)
     token_count = ndb.IntegerProperty(indexed=False)
     currency = ndb.StringProperty(indexed=False, choices=FULL_CURRENCY_NAMES.keys())
